requirements.txt
python-docx
selenium
requests
#pip install -r requirements.txt
